final version

+++ ocs2_test_tools/ocs2_qp_solver/test/testQpSolver.cpp
@@ -9,12 +9,12 @@
 
 class QpSolverTest : public testing::Test {
  protected:
-  static constexpr int N_ = 5;
-  static constexpr int nx_ = 4;
-  static constexpr int nu_ = 3;
-  static constexpr int nc_ = 2;
-  static constexpr int numDecisionVariables = N_ * (nx_ + nu_) + nx_;
-  static constexpr int numConstraints = (N_ + 1) * (nx_ + nc_);
+  static constexpr size_t N_ = 5;
+  static constexpr size_t nx_ = 4;
+  static constexpr size_t nu_ = 3;
+  static constexpr size_t nc_ = 2;
+  static constexpr size_t numDecisionVariables = N_ * (nx_ + nu_) + nx_;
+  static constexpr size_t numConstraints = (N_ + 1) * (nx_ + nc_);
 
   QpSolverTest() {
     srand(0);
@@ -34,12 +34,12 @@ class QpSolverTest : public testing::Test {
   ocs2::dynamic_vector_t dualSolution;
 };
 
-constexpr int QpSolverTest::N_;
-constexpr int QpSolverTest::nx_;
-constexpr int QpSolverTest::nu_;
-constexpr int QpSolverTest::nc_;
-constexpr int QpSolverTest::numDecisionVariables;
-constexpr int QpSolverTest::numConstraints;
+constexpr size_t QpSolverTest::N_;
+constexpr size_t QpSolverTest::nx_;
+constexpr size_t QpSolverTest::nu_;
+constexpr size_t QpSolverTest::nc_;
+constexpr size_t QpSolverTest::numDecisionVariables;
+constexpr size_t QpSolverTest::numConstraints;
 
 TEST_F(QpSolverTest, constraintSatisfaction) {
   ASSERT_TRUE(constraints.f.isApprox(-constraints.dfdx * primalSolution));
